💄 edited header's style
