+++ src/components/header/header.css
@@ -1,0 +1,26 @@
+.header {
+  position: fixed;
+  left: 0;
+  top: 0;
+  z-index: 9999;
+  width: 100%;
+  height: 137px;
+  background-color: #ffffff;
+}
+
+.header .contents {
+  display: flex;
+  width: 96%;
+  max-width: 1440px;
+  height: 100%;
+  margin: auto;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.header .img {
+  width: 281px;
+  height: 118px;
+  object-fit: cover;
+  outline: none;
+}
