+++ src/components/header/menuBtn.jsx
@@ -12,9 +12,7 @@ const MenuBtnSet = ({ text1, url1, text2, url2, text3, url3 }) => {
   return (
     <div className="menuBtnSet">
       <MenuBtn text={text1} url={url1} />
-      &nbsp;
       <MenuBtn text={text2} url={url2} />
-      &nbsp;
       <MenuBtn text={text3} url={url3} />
     </div>
   );
